refactor(models): extract shared Base
