+++ infrastructure/models/__init__.py
@@ -1,5 +1,9 @@
-import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("infrastructure > models")
+import logging
+
+logging.basicConfig(level=logging.DEBUG)
+logging.debug("infrastructure > models")
+from .base import Base
 from .company_model import CompanyModel
 from .nsd_model import NSDModel
 
-__all__ = ["CompanyModel", "NSDModel"]
+__all__ = ["Base", "CompanyModel", "NSDModel"]
