+++ infrastructure/models/company_model.py
@@ -1,25 +1,30 @@
 from __future__ import annotations
-import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("infrastructure > models > company_model")
+
+import logging
+
+logging.basicConfig(level=logging.DEBUG)
+logging.debug("infrastructure > models > company_model")
 
 import json
 from datetime import datetime
 from typing import List, Optional
 
 from sqlalchemy import Boolean, DateTime
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
+from sqlalchemy.orm import Mapped, mapped_column
 
 from domain.dto.company_dto import CompanyDTO
 from domain.dto.raw_company_dto import CodeDTO, CompanyRawDTO
 
-
-class Base(DeclarativeBase):
-    import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("company_model.Base")
-    pass
+from .base import Base
 
 
 class CompanyModel(Base):
     """ORM adapter for the ``tbl_company`` table."""
-    import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("company_model.CompanyModel(Base)")
+
+    import logging
+
+    logging.basicConfig(level=logging.DEBUG)
+    logging.debug("company_model.CompanyModel(Base)")
 
     __tablename__ = "tbl_company"
 
@@ -67,7 +72,10 @@ class CompanyModel(Base):
 
     @staticmethod
     def from_dto(dto: CompanyRawDTO | CompanyDTO) -> "CompanyModel":
-        import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("CompanyModel.from_dto()")
+        import logging
+
+        logging.basicConfig(level=logging.DEBUG)
+        logging.debug("CompanyModel.from_dto()")
         """Convert a ``CompanyRawDTO`` or ``CompanyDTO`` into ``CompanyModel``."""
 
         def attr(name: str):
@@ -121,7 +129,10 @@ class CompanyModel(Base):
         )
 
     def to_dto(self) -> CompanyRawDTO:
-        import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("CompanyModel.to_dto()")
+        import logging
+
+        logging.basicConfig(level=logging.DEBUG)
+        logging.debug("CompanyModel.to_dto()")
         """Reconstruct a :class:`CompanyRawDTO` from this model."""
 
         ticker_codes: List[str] = (
