+++ infrastructure/repositories/company_repository.py
@@ -1,5 +1,9 @@
 from __future__ import annotations
-import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("infrastructure > company_repository")
+
+import logging
+
+logging.basicConfig(level=logging.DEBUG)
+logging.debug("infrastructure > company_repository")
 
 from typing import List
 
@@ -10,7 +14,8 @@ from domain.dto.company_dto import CompanyDTO
 from domain.ports import CompanyRepositoryPort
 from infrastructure.config import Config
 from infrastructure.logging import Logger
-from infrastructure.models.company_model import Base, CompanyModel
+from infrastructure.models.base import Base
+from infrastructure.models.company_model import CompanyModel
 from infrastructure.repositories.base_repository import BaseRepository
 
 
@@ -19,13 +24,24 @@ class SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort)
     Concrete implementation of BaseRepository for CompanyDTO,
     using SQLite and SQLAlchemy for persistence.
     """
-    import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("company_repository.SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort)")
+
+    import logging
+
+    logging.basicConfig(level=logging.DEBUG)
+    logging.debug(
+        "company_repository.SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort)"
+    )
 
     def __init__(self, config: Config, logger: Logger):
         """
         Initializes the SQLite database connection and ensures table creation.
         """
-        import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).__init__")
+        import logging
+
+        logging.basicConfig(level=logging.DEBUG)
+        logging.debug(
+            "SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).__init__"
+        )
         self.config = config
         self.logger = logger
         self.logger.log("Start SQLiteCompanyRepository", level="info")
@@ -45,7 +61,12 @@ class SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort)
 
         :param items: List of CompanyDTO instances to persist
         """
-        import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).save_all()")
+        import logging
+
+        logging.basicConfig(level=logging.DEBUG)
+        logging.debug(
+            "SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).save_all()"
+        )
         session = self.Session()
         try:
             models = [CompanyModel.from_dto(dto) for dto in items]
@@ -71,7 +92,12 @@ class SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort)
 
         :return: A list of CompanyDTOs
         """
-        import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).get_all()")
+        import logging
+
+        logging.basicConfig(level=logging.DEBUG)
+        logging.debug(
+            "SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).get_all()"
+        )
         session = self.Session()
         try:
             results = session.query(CompanyModel).all()
@@ -86,7 +112,12 @@ class SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort)
         :param identifier: CVM code to verify
         :return: True if the company exists, False otherwise
         """
-        import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).has_item()")
+        import logging
+
+        logging.basicConfig(level=logging.DEBUG)
+        logging.debug(
+            "SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).has_item()"
+        )
         session = self.Session()
         try:
             return (
@@ -104,7 +135,12 @@ class SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort)
         :return: A CompanyDTO representing the retrieved company
         :raises ValueError: If no company is found
         """
-        import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).get_by_id()")
+        import logging
+
+        logging.basicConfig(level=logging.DEBUG)
+        logging.debug(
+            "SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).get_by_id()"
+        )
         session = self.Session()
         try:
             obj = session.query(CompanyModel).filter_by(cvm_code=id).first()
@@ -121,7 +157,12 @@ class SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort)
 
         :return: Conjunto de códigos CVM únicos.
         """
-        import logging; logging.basicConfig(level=logging.DEBUG); logging.debug("SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).get_all_primary_keys()")
+        import logging
+
+        logging.basicConfig(level=logging.DEBUG)
+        logging.debug(
+            "SQLiteCompanyRepository(BaseRepository[CompanyDTO], CompanyRepositoryPort).get_all_primary_keys()"
+        )
         session = self.Session()
         try:
             results = session.query(CompanyModel.cvm_code).distinct().all()
