+++ tests/infrastructure/test_company_repository.py
@@ -1,7 +1,7 @@
 from sqlalchemy import text
 
 from domain.dto.company_dto import CompanyDTO
-from infrastructure.models.company_model import Base
+from infrastructure.models.base import Base
 from infrastructure.repositories.company_repository import SQLiteCompanyRepository
 from tests.conftest import DummyConfig, DummyLogger
